[fix] makeMoim Errors
