+++ app/src/main/res/font/font.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="utf-8"?>
+<font-family xmlns:tools="http://schemas.android.com/tools"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <font
+        android:font="@font/notosans_kr_default"
+        android:fontStyle="normal"
+        android:fontWeight="400"
+        tools:targetApi="o" />
+
+    <font
+        android:font="@font/notosans_kr_bold"
+        android:fontStyle="normal"
+        android:fontWeight="400"
+        tools:targetApi="o"/>
+
+    <font
+        android:font="@font/notosans_kr_light"
+        android:fontStyle="normal"
+        android:fontWeight="400"
+        tools:targetApi="o"/>
+
+    <font
+        android:font="@font/notosans_kr_medium"
+        android:fontStyle="normal"
+        android:fontWeight="400"
+        tools:targetApi="o"/>
+
+    <font
+        android:font="@font/notosans_kr_regular"
+        android:fontStyle="normal"
+        android:fontWeight="400"
+        tools:targetApi="o"/>
+
+    <font
+        android:font="@font/notosans_kr_thin"
+        android:fontStyle="normal"
+        android:fontWeight="400"
+        tools:targetApi="o"/>
+
+</font-family>
